+++ internal/prometheus/recording_rules.go
@@ -186,11 +186,13 @@ count_over_time({{.metric}}{{.filter}}[{{.window}}])
 	return &rulefmt.Rule{
 		Record: slo.GetSLIErrorMetric(window),
 		Expr:   b.String(),
-		// The SLO labels will be obtained from the source SLI recording rule.
-		// We only need to set the window.
-		Labels: map[string]string{
-			sloWindowLabelName: strWindow,
-		},
+		Labels: mergeLabels(
+			slo.GetSLOIDPromLabels(),
+			map[string]string{
+				sloWindowLabelName: strWindow,
+			},
+			slo.Labels,
+		),
 	}, nil
 }